Extracted CLI to its own project
