+++ ScrapPackedExplorerCli/Properties/launchSettings.json
@@ -3,9 +3,6 @@
     "ScrapPackedExplorerCli": {
       "commandName": "Project",
       "commandLineArgs": "--help"
-    },
-    "ScrapPackedExplorerCliAsGUI": {
-      "commandName": "Project"
     }
   }
 }
