+++ ScrapPackedExplorerCombined/Properties/launchSettings.json
@@ -1,0 +1,11 @@
+{
+  "profiles": {
+    "ScrapPackedExplorerAsGUI": {
+      "commandName": "Project"
+    },
+    "ScrapPackedExplorerAsCLI": {
+      "commandName": "Project",
+      "commandLineArgs": "--help"
+    }
+  }
+}
