+++ ScrapPackedLibrary/ScrapPackedLibrary.csproj
@@ -1,7 +1,7 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>
-    <TargetFramework>net6.0-windows</TargetFramework>
+    <TargetFramework>net6.0</TargetFramework>
     <RootNamespace>ch.romibi.Scrap.Packed.PackerLib</RootNamespace>
     <PackageId>ScrapPackedLibrary</PackageId>
     <Authors>romibi, Strongleong</Authors>
